support trace models to onnx or jit model
